Move IPerfSignalReport.
