+++ packages/runtime/container-runtime/src/connectionTelemetry.ts
@@ -457,57 +457,6 @@ class OpPerfTelemetry {
 		}
 	}
 }
-export interface IPerfSignalReport {
-	/**
-	 * Identifier to track broadcast signals being submitted in order to
-	 * allow collection of data around the roundtrip of signal messages.
-	 */
-	broadcastSignalSequenceNumber: number;
-
-	/**
-	 * Accumulates the total number of broadcast signals sent during the current signal latency measurement window.
-	 * This value represents the total number of signals sent since the latency measurement began and is used
-	 * logged in telemetry when the latency measurement completes.
-	 */
-	totalSignalsSentInLatencyWindow: number;
-
-	/**
-	 * Counts the number of broadcast signals sent since the last latency measurement was initiated.
-	 * This counter increments with each broadcast signal sent. When a new latency measurement starts,
-	 * this counter is added to `totalSignalsSentInLatencyWindow` and then reset to zero.
-	 */
-	signalsSentSinceLastLatencyMeasurement: number;
-
-	/**
-	 * Number of signals that were expected but not received.
-	 */
-	signalsLost: number;
-
-	/**
-	 * Number of signals received out of order/non-sequentially.
-	 */
-	signalsOutOfOrder: number;
-
-	/**
-	 * Timestamp before submitting the signal we will trace.
-	 */
-	signalTimestamp: number;
-
-	/**
-	 * Signal we will trace for roundtrip latency.
-	 */
-	roundTripSignalSequenceNumber: number | undefined;
-
-	/**
-	 * Next expected signal sequence number to be received.
-	 */
-	trackingSignalSequenceNumber: number | undefined;
-
-	/**
-	 * Inclusive lower bound of signal monitoring window.
-	 */
-	minimumTrackingSignalSequenceNumber: number | undefined;
-}
 
 /**
  * Starts monitoring and generation of telemetry related to op performance.
